WebStorage/tests/submissions/Infraware/storage_session_setitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_session_setItem</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('sessionStorage' in window) && window.sessionStorage !== null){
			sessionStorage.clear();

			try {
				sessionStorage.setItem("name", "user1");
			} catch (e) {
				test(function()
				{
					assert_unreached("setItem method is failed.");
				}, "setItem method test")
			}

			test(function() {
				assert_not_equals(sessionStorage.length, 0, "sessionStorage.setItem")
			}, "setItem method test")

		} else {
			test(function()
			{
				assert_unreached("sessionStorage is not supported.");
			}, "sessionStorage support test")
		}	
	</script>

 </body>
</html>
